refactor(contract): restructure console contracts with nested billing module
Reorganize oRPC console contracts to follow best practices from the
architecture guide. Move from flat structure to nested modules for
better scalability as more APIs are migrated.

Changes:
- Split console.ts into console/system.ts and console/billing.ts
- Nest billing contracts under `billing` key in router
- Rename billingUrlContract to invoicesContract for clarity
- Update all call sites to use new nested paths
