+++ web/contract/console/billing.ts
@@ -1,16 +1,7 @@
-import type { SystemFeatures } from '@/types/feature'
 import { type } from '@orpc/contract'
-import { base } from './base'
+import { base } from '../base'
 
-export const systemFeaturesContract = base
-  .route({
-    path: '/system-features',
-    method: 'GET',
-  })
-  .input(type<unknown>())
-  .output(type<SystemFeatures>())
-
-export const billingUrlContract = base
+export const invoicesContract = base
   .route({
     path: '/billing/invoices',
     method: 'GET',
